refactor: project name, artifacts, group 명 변경
