--- src/main/java/com/example/demo/common/aop/A.java
@@ -1,4 +1,0 @@
-package com.example.demo.common.aop;
-
-public class A {
-}
