--- src/main/java/com/example/demo/common/jwt/J.java
@@ -1,4 +1,0 @@
-package com.example.demo.common.jwt;
-
-public class J {
-}
