--- src/main/java/com/example/demo/common/util/U.java
@@ -1,4 +1,0 @@
-package com.example.demo.common.util;
-
-public class U {
-}
