+++ src/main/java/com/odiga/fiesta/DemoApplication.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.odiga.fiesta;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
