+++ src/main/java/com/odiga/fiesta/common/aop/A.java
@@ -1,0 +1,4 @@
+package com.odiga.fiesta.common.aop;
+
+public class A {
+}
