+++ src/main/java/com/odiga/fiesta/common/config/JpaAuditingConfig.java
@@ -1,4 +1,4 @@
-package com.example.demo.common.config;
+package com.odiga.fiesta.common.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
