+++ src/main/java/com/odiga/fiesta/common/domain/BaseEntity.java
@@ -1,4 +1,4 @@
-package com.example.demo.common.domain;
+package com.odiga.fiesta.common.domain;
 
 import java.time.LocalDateTime;
