+++ src/main/java/com/odiga/fiesta/common/error/ErrorResponse.java
@@ -1,4 +1,4 @@
-package com.example.demo.common.error;
+package com.odiga.fiesta.common.error;
 
 import lombok.AccessLevel;
 import lombok.Builder;
