+++ src/main/java/com/odiga/fiesta/common/error/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package com.example.demo.common.error;
+package com.odiga.fiesta.common.error;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.http.converter.HttpMessageNotReadableException;
