+++ src/main/java/com/odiga/fiesta/common/jwt/J.java
@@ -1,0 +1,4 @@
+package com.odiga.fiesta.common.jwt;
+
+public class J {
+}
