+++ src/main/java/com/odiga/fiesta/common/util/U.java
@@ -1,0 +1,4 @@
+package com.odiga.fiesta.common.util;
+
+public class U {
+}
