+++ src/main/java/com/odiga/fiesta/hello/controller/HelloController.java
@@ -1,10 +1,10 @@
-package com.example.demo.hello.controller;
+package com.odiga.fiesta.hello.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.example.demo.common.ApiResponse;
+import com.odiga.fiesta.common.ApiResponse;
 
 @RestController
 public class HelloController {
