+++ src/test/java/com/odiga/fiesta/ControllerTestSupport.java
@@ -1,10 +1,10 @@
-package com.example.demo;
+package com.odiga.fiesta;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.test.web.servlet.MockMvc;
 
-import com.example.demo.hello.controller.HelloController;
+import com.odiga.fiesta.hello.controller.HelloController;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 @WebMvcTest(controllers = {
