+++ src/test/java/com/odiga/fiesta/DemoApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.odiga.fiesta;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
